Renamed files to better reflect new API
